Streamline lots of code and remove old kernel
